Moved common code from CLIP and GPT2 tokenizers into separate file
