+++ operators/tokenizer/tokenizers.cc
@@ -3,9 +3,6 @@
 
 #ifdef ENABLE_GPT2_TOKENIZER
 #include "gpt2_tokenizer.hpp"
-#endif
-
-#ifdef ENABLE_CLIP_TOKENIZER
 #include "clip_tokenizer.hpp"
 #endif
 
@@ -33,9 +30,6 @@ FxLoadCustomOpFactory LoadCustomOpClasses_Tokenizer = LoadCustomOpClasses<
     CustomOpClassBegin
 #ifdef ENABLE_GPT2_TOKENIZER
     , CustomOpBpeTokenizer
-#endif
-
-#ifdef ENABLE_CLIP_TOKENIZER
     , CustomOpClipBpeTokenizer
 #endif
